pyproject.toml
[tool.poetry]
name = "video-merge"
version = "0.1.0"
description = ""
authors = ["MohamedMagdy097 <zalatanback5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
nest-asyncio = "^1.5.9"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
retry = "^0.9.2"
aiofiles = "^24.1.0"
tenacity = "^9.0.0"
aiohttp = "^3.11.11"
supabase = "^2.11.0"
asyncer = "^0.0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "scripts.index:start"
dev = "scripts.index:dev"
dev_win = "scripts.index:dev_win"